📈 perf(client): devide search bar from Navbar
Navbar안에 있던 searchbox를 분리하였습니다
