+++ packages/client/src/App.tsx
@@ -2,10 +2,12 @@ import './App.css';
 import { Route, Router, Routes } from 'react-router-dom';
 import LandingPage from './pages/LandingPage';
 import ProfilePage from './pages/ProfilePage';
+import ChatroomPage from './pages/ChatroomPage';
 
 function App() {
   return (
     <Routes>
+      <Route path="/chatRoom" element={<ChatroomPage />} />
       <Route path="/profile" element={<ProfilePage />} />
       <Route path="/" element={<LandingPage />} />
     </Routes>
