+++ packages/client/src/components/bar/NavigationBar.tsx
@@ -14,44 +14,7 @@ import SportsEsportsIcon from '@mui/icons-material/SportsEsports';
 import GameIconButton from '../button/icon/GameIconButton';
 import LogoutIconButton from '../button/icon/LogoutIconButton';
 import { flexbox } from '@mui/system';
-
-const Search = styled('div')(({ theme }) => ({
-  position: 'relative',
-  borderRadius: theme.shape.borderRadius,
-  backgroundColor: alpha(theme.palette.common.white, 0.15),
-  '&:hover': {
-    backgroundColor: alpha(theme.palette.common.white, 0.25),
-  },
-  marginLeft: 'auto',
-  width: '100%',
-}));
-
-const SearchIconWrapper = styled('div')(({ theme }) => ({
-  padding: theme.spacing(0, 2),
-  height: '100%',
-  position: 'absolute',
-  pointerEvents: 'none',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-}));
-
-const StyledInputBase = styled(InputBase)(({ theme }) => ({
-  color: 'inherit',
-  '& .MuiInputBase-input': {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
-    transition: theme.transitions.create('width'),
-    width: '100%',
-    [theme.breakpoints.up('sm')]: {
-      width: '12ch',
-      '&:focus': {
-        width: '20ch',
-      },
-    },
-  },
-}));
+import SearchBox from '../input/SearchBox';
 
 const RightLayout = styled('section')(({ theme }) => ({
   marginLeft: 'auto',
@@ -68,15 +31,7 @@ export default function NavigationBar() {
           <GameIconButton />
           <ChatIconButton />
           <RightLayout>
-            <Search>
-              <SearchIconWrapper>
-                <SearchIcon />
-              </SearchIconWrapper>
-              <StyledInputBase
-                placeholder="유저찾기…"
-                inputProps={{ 'aria-label': 'search' }}
-              />
-            </Search>
+            <SearchBox />
             <LogoutIconButton />
           </RightLayout>
         </Toolbar>
