+++ packages/client/src/components/button/icon/LogoutButton.tsx
@@ -1,0 +1,12 @@
+import LogoutIcon from '@mui/icons-material/Logout';
+import { IconButton } from '@mui/material';
+
+function LogoutIconButton() {
+  return (
+    <IconButton aria-label="game" size="large">
+      <LogoutIcon fontSize="inherit" />
+    </IconButton>
+  );
+}
+
+export default LogoutIconButton;
